+++ app/controllers/dorsale/customer_vault/people_controller.rb
@@ -213,6 +213,7 @@ class Dorsale::CustomerVault::PeopleController < ::Dorsale::CustomerVault::Appli
       :author => current_user,
       :person => @person,
       :action => action,
+      :text   => ::Dorsale::CustomerVault::Event.t("action.#{action}")
     )
   end